Moving older, unused autonomous programs into archive directory
