+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/archive/AutoBlueLameProgram.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.archive;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
@@ -6,8 +6,8 @@ import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-@Autonomous(name = "Epik Autonomous Red Forward", group = "Linear Opmode")
-public class AutoRedLameProgram extends LinearOpMode {
+@Autonomous(name = "Epik Autonomous Blue Forward", group = "Linear Opmode")
+public class AutoBlueLameProgram extends LinearOpMode {
     private DcMotor leftFrontDrive = null;
     private DcMotor leftBackDrive = null;
     private DcMotor rightFrontDrive = null;
