+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/archive/AutoRed.java
@@ -1,13 +1,14 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.archive;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-@Autonomous(name = "Epik Autonomous Blue", group = "Linear Opmode")
-public class AutoBlue extends LinearOpMode {
+@Autonomous(name = "Epik Autonomous Red", group = "Linear Opmode")
+public class AutoRed extends LinearOpMode {
     private DcMotor leftFrontDrive = null;
     private DcMotor leftBackDrive = null;
     private DcMotor rightFrontDrive = null;
@@ -59,10 +60,10 @@ public class AutoBlue extends LinearOpMode {
 
             runtime.reset();
             while (runtime.seconds() < 1.7) {
-                leftFrontDrive.setPower(-0.5);
-                rightFrontDrive.setPower(0.5);
-                rightBackDrive.setPower(-0.5);
-                leftBackDrive.setPower(0.5);
+                leftFrontDrive.setPower(0.5);
+                rightFrontDrive.setPower(-0.5);
+                rightBackDrive.setPower(0.5);
+                leftBackDrive.setPower(-0.5);
             }
 
             leftFrontDrive.setPower(0);
