+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/archive/AutoRedLameProgram.java
@@ -1,14 +1,13 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.archive;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-@Autonomous(name = "Epik Autonomous Red", group = "Linear Opmode")
-public class AutoRed extends LinearOpMode {
+@Autonomous(name = "Epik Autonomous Red Forward", group = "Linear Opmode")
+public class AutoRedLameProgram extends LinearOpMode {
     private DcMotor leftFrontDrive = null;
     private DcMotor leftBackDrive = null;
     private DcMotor rightFrontDrive = null;
@@ -43,27 +42,6 @@ public class AutoRed extends LinearOpMode {
                 rightFrontDrive.setPower(0.5);
                 rightBackDrive.setPower(0.5);
                 leftBackDrive.setPower(0.5);
-            }
-
-            leftFrontDrive.setPower(0);
-            rightFrontDrive.setPower(0);
-            rightBackDrive.setPower(0);
-            leftBackDrive.setPower(0);
-
-            runtime.reset();
-            while (runtime.seconds() < 0.1) {
-                leftFrontDrive.setPower(-0.5);
-                rightFrontDrive.setPower(-0.5);
-                rightBackDrive.setPower(-0.5);
-                leftBackDrive.setPower(-0.5);
-            }
-
-            runtime.reset();
-            while (runtime.seconds() < 1.7) {
-                leftFrontDrive.setPower(0.5);
-                rightFrontDrive.setPower(-0.5);
-                rightBackDrive.setPower(0.5);
-                leftBackDrive.setPower(-0.5);
             }
 
             leftFrontDrive.setPower(0);
